% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotowanie_danych.R
\name{.analizuj_skladnia_smm}
\alias{.analizuj_skladnia_smm}
\title{Wewnetrzny parser modelu badawczego}
\usage{
.analizuj_skladnia_smm(skladnia)
}
\arguments{
\item{skladnia}{Ciag znakow ze skladnia modelu.}
}
\value{
Struktura listowa z mapowaniem kryteriow na zmienne surowe.
}
\description{
Dekoduje definicje kryteriow MCDA podane w skladni zblizonej do lavaan (np. "Kryterium =~ zm1 + zm2").
}
\keyword{internal}
