% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wizualizacja_rankingu.R
\name{.motyw_smm}
\alias{.motyw_smm}
\title{Wewnetrzny motyw graficzny dla wykresow SMM}
\usage{
.motyw_smm()
}
\description{
Ujednolicony, minimalistyczny styl wykresow dla calego pakietu.
}
\keyword{internal}
