% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotowanie_danych.R
\name{.skaluj_do_1_9}
\alias{.skaluj_do_1_9}
\title{Wewnetrzny skaler do przedzialu 1-9}
\usage{
.skaluj_do_1_9(x)
}
\arguments{
\item{x}{Wektor numeryczny.}
}
\value{
Wektor sformatowany do przedzialu [1, 9].
}
\description{
Przeksztalca wektor wartosci liczbowych na jednolita skale od 1 do 9.
}
\keyword{internal}
