% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorytm_mcda.R
\name{oblicz_meta_ranking_smm}
\alias{oblicz_meta_ranking_smm}
\title{Meta-Ranking platform SMM}
\usage{
oblicz_meta_ranking_smm(
  tablica_wynikow,
  wagi_smm,
  kierunki = NULL,
  v = 0.5,
  lambda = 0.5
)
}
\arguments{
\item{tablica_wynikow}{Macierz decyzyjna (alternatywy x kryteria).}

\item{wagi_smm}{Wektor wag kryteriów sumujący się do 1.}

\item{kierunki}{Wektor znakowy kierunków kryteriów ("max" lub "min").}

\item{v}{Parametr dla VIKOR (domyślnie 0.5).}

\item{lambda}{Parametr dla WASPAS (domyślnie 0.5).}
}
\value{
Obiekt klasy \code{meta_smm_wynik} zawierający tabelę porównawczą i macierz korelacji.
}
\description{
Agreguje wyniki z metod TOPSIS, VIKOR i WASPAS w jeden spójny meta-ranking konsensusu.
}
