% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorytm_mcda.R
\name{oblicz_vikor_smm}
\alias{oblicz_vikor_smm}
\title{Ranking SMM metodą VIKOR}
\usage{
oblicz_vikor_smm(tablica_wynikow, wagi_smm, kierunki = NULL, v = 0.5)
}
\arguments{
\item{tablica_wynikow}{Macierz decyzyjna (alternatywy x kryteria).}

\item{wagi_smm}{Wektor wag kryteriów sumujący się do 1.}

\item{kierunki}{Wektor znakowy określający kierunki kryteriów ("max" lub "min").}

\item{v}{Parametr wagowy strategii "większości kryteriów" (zwykle 0.5).}
}
\value{
Obiekt klasy \code{vikor_smm_wynik} z tabelą rankingu i detalami.
}
\description{
Wykonuje procedurę kompromisowego rankingu VIKOR dla danych ostrych.
}
