% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metody_wazenia.R
\name{oblicz_wagi_entropia}
\alias{oblicz_wagi_entropia}
\title{Wagi Entropii Shannona}
\usage{
oblicz_wagi_entropia(macierz)
}
\arguments{
\item{macierz}{Macierz decyzyjna (alternatywy x kryteria).}
}
\value{
Wektor wag kryteriów sumujący się do 1.
}
\description{
Oblicza obiektywne wagi kryteriów na podstawie miary entropii (rozproszenia informacji).
}
