% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metoda_wag_bwm.R
\name{oblicz_wagi_smm_bwm}
\alias{oblicz_wagi_smm_bwm}
\title{Wyznaczanie wag kryteriów metodą Best-Worst (BWM)}
\usage{
oblicz_wagi_smm_bwm(najlepszy_indeks, najgorszy_indeks, ile_kryteriow)
}
\arguments{
\item{najlepszy_indeks}{Indeks kryterium uznanego za najważniejsze}

\item{najgorszy_indeks}{Indeks kryterium uznanego za najmniej ważne}

\item{ile_kryteriow}{Całkowita liczba kryteriów w analizie}
}
\value{
wagi_koncowe Wektor wag, których suma wynosi 1
}
\description{
Funkcja oblicza wagi dla kryteriów SMM na podstawie wyboru najlepszego i najgorszego elementu.
}
