% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metody_wazenia.R
\name{oblicz_wagi_std_dev}
\alias{oblicz_wagi_std_dev}
\title{Wagi Odchylenia Standardowego}
\usage{
oblicz_wagi_std_dev(macierz, kierunki = NULL)
}
\arguments{
\item{macierz}{Macierz decyzyjna (alternatywy x kryteria).}

\item{kierunki}{Opcjonalny wektor kierunków kryteriów ("max" lub "min"). Domyślnie wszystkie są traktowane jako "max".}
}
\value{
Wektor wag kryteriów sumujący się do 1.
}
\description{
Wyznacza obiektywne wagi kryteriów na podstawie odchylenia standardowego po normalizacji Min-Max.
}
