% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorytm_mcda.R
\name{oblicz_waspas_smm}
\alias{oblicz_waspas_smm}
\title{Ranking SMM metodą WASPAS}
\usage{
oblicz_waspas_smm(tablica_wynikow, wagi_smm, kierunki = NULL, lambda = 0.5)
}
\arguments{
\item{tablica_wynikow}{Macierz decyzyjna (alternatywy x kryteria).}

\item{wagi_smm}{Wektor wag kryteriów sumujący się do 1.}

\item{kierunki}{Wektor znakowy określający kierunki kryteriów ("max" lub "min").}

\item{lambda}{Parametr równowagi między modelem sumarycznym WSM a iloczynowym WPM. Domyślnie 0.5.}
}
\value{
Obiekt klasy \code{waspas_smm_wynik} z tabelą rankingu.
}
\description{
Wykonuje procedurę rankingu WASPAS (Weighted Aggregated Sum Product Assessment).
}
