% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wizualizacja_rankingu.R
\name{plot.topsis_smm_wynik}
\alias{plot.topsis_smm_wynik}
\title{Mapa Efektywnosci TOPSIS}
\usage{
\method{plot}{topsis_smm_wynik}(x, ...)
}
\arguments{
\item{x}{Obiekt klasy \code{topsis_smm_wynik}.}

\item{...}{Dodatkowe argumenty (ignorowane).}
}
\description{
Wizualizuje odleglosci alternatyw od idealu i anty-idealu.
Os X: Dystans od Najgorszego (D-). Os Y: Dystans do Najlepszego (D+).
Najlepsze platformy znajduja sie w prawym dolnym rogu.
}
