% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wizualizacja_rankingu.R
\name{plot.vikor_smm_wynik}
\alias{plot.vikor_smm_wynik}
\title{Mapa Strategiczna VIKOR}
\usage{
\method{plot}{vikor_smm_wynik}(x, ...)
}
\arguments{
\item{x}{Obiekt klasy \code{vikor_smm_wynik}.}

\item{...}{Dodatkowe argumenty (ignorowane).}
}
\description{
Pokazuje kompromis miedzy maksymalizacja uzytecznosci grupowej (os X)
a minimalizacja indywidualnego zalu (os Y).
}
