% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wizualizacja_rankingu.R
\name{plot.waspas_smm_wynik}
\alias{plot.waspas_smm_wynik}
\title{Mapa Balansu WASPAS}
\usage{
\method{plot}{waspas_smm_wynik}(x, ...)
}
\arguments{
\item{x}{Obiekt klasy \code{waspas_smm_wynik}.}

\item{...}{Dodatkowe argumenty (ignorowane).}
}
\description{
Prezentuje relacje miedzy wynikiem sumarycznym (WSM) a iloczynowym (WPM).
Najlepsze platformy leza w prawym gornym rogu.
}
