% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wizualizacja_rankingu.R
\name{rysuj_ranking_smm}
\alias{rysuj_ranking_smm}
\title{Rysowanie rankingu SMM (funkcja kompatybilnosci)}
\usage{
rysuj_ranking_smm(macierz_danych, wyniki_rankingu)
}
\arguments{
\item{macierz_danych}{Macierz z danymi.}

\item{wyniki_rankingu}{Wyniki rankingu (np. z TOPSIS).}
}
\description{
Tradycyjny rysunek zachowany dla kompatybilnosci wstecznej, wykorzystujacy
nowa logike graficzna ggplot2.
}
