% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smm_bubble_plot.R
\name{smm_bubble_plot}
\alias{smm_bubble_plot}
\title{Wykres bubble dla wyników MCDA}
\usage{
smm_bubble_plot(scores, weights)
}
\arguments{
\item{scores}{data.frame lub macierz z ocenami kryteriów}

\item{weights}{wektor wag kryteriów}
}
\description{
Wykres bubble dla wyników MCDA
}
