% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smm_dane_eksperckie}
\alias{smm_dane_eksperckie}
\title{Dane eksperckie dla platform SMM}
\format{
Macierz o 3 wierszach i 3 kolumnach:
\describe{
  \item{Zasieg}{Ocena zasięgu}
  \item{Zaangazowanie}{Ocena zaangażowania}
  \item{Koszt}{Ocena kosztu}
}
}
\source{
Dane eksperckie z oryginalnego pakietu.
}
\usage{
smm_dane_eksperckie
}
\description{
Klasyczna macierz ocen eksperckich dla platform TikTok, Instagram i YouTube
pod kątem Zasięgu, Zaangażowania i Kosztów.
}
\keyword{datasets}
