% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smm_dane_surowe}
\alias{smm_dane_surowe}
\title{Przykładowe surowe dane marketingowe dla platform SMM}
\format{
Ramka danych o 120 wierszach i 14 kolumnach:
\describe{
  \item{Kampania_ID}{Unikalny identyfikator kampanii (1-120)}
  \item{Platforma}{Platforma społecznościowa (TikTok, Instagram, YouTube, Facebook, LinkedIn, Pinterest)}
  \item{zasieg_surowy}{Surowy zasięg kampanii (wyświetlenia)}
  \item{liczba_klikniec}{Liczba kliknięć (z błędami 999)}
  \item{wspolczynnik_konwersji}{Współczynnik konwersji w procentach (z brakami NA)}
  \item{polubienia}{Liczba polubień}
  \item{komentarze}{Liczba komentarzy}
  \item{udostepnienia}{Liczba udostępnień}
  \item{koszt_kampanii}{Koszt kampanii w PLN}
  \item{cpc}{Koszt za kliknięcie (Cost Per Click) w PLN}
  \item{cpm}{Koszt za tysiąc wyświetleń (Cost Per Mille) w PLN}
  \item{satysfakcja_klienta}{Zadowolenie klienta z kampanii (Likert 1-5, z błędami 999)}
  \item{latwosc_obslugi}{Łatwość konfiguracji/obsługi platformy (Likert 1-5)}
  \item{jakosc_wspolpracy}{Jakość współpracy z influencerami/wsparciem (Likert 1-5, z brakami NA)}
}
}
\source{
Syntetyczne generowanie w R (DGP).
}
\usage{
smm_dane_surowe
}
\description{
Syntetyczny zbiór danych zawierający wyniki 120 kampanii marketingowych
przeprowadzonych na 6 platformach społecznościowych. Zbiór zawiera celowo
wprowadzone błędy (np. 999) oraz braki danych (NA) do celów demonstracyjnych.
}
\keyword{datasets}
