% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smm_mcdm_rank.R
\name{smm_rank_score}
\alias{smm_rank_score}
\title{Ranking platform spolecznosciowych metoda MCDA}
\usage{
smm_rank_score(scores, weights)
}
\arguments{
\item{scores}{Macierz lub data.frame z ocenami alternatyw (wiersze to platformy, kolumny to kryteria)}

\item{weights}{Wektor wag dla kryteriow (suma wag powinna wynosic 1)}
}
\value{
Data.frame z wynikami rankingu i punktacja
}
\description{
Ranking platform spolecznosciowych metoda MCDA
}
