% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dane_marketingowe.R
\name{stworz_macierz_smm}
\alias{stworz_macierz_smm}
\title{Generowanie macierzy danych dla platform SMM}
\usage{
stworz_macierz_smm()
}
\value{
macierz_smm Macierz z danymi marketingowymi (Zasięg, Zaangażowanie, Koszt).
}
\description{
Funkcja tworzy profesjonalną macierz decyzyjną z ocenami dla platform TikTok, Instagram i YouTube.
}
