% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabela_apa.R
\name{tabela_apa}
\alias{tabela_apa}
\title{Generowanie Tabeli APA dla SMMRankR}
\usage{
tabela_apa(x, tytul = NULL)
}
\arguments{
\item{x}{Obiekt wynikowy z funkcji pakietu (np. \code{topsis_smm_wynik}).}

\item{tytul}{Opcjonalny tytul tabeli.}
}
\value{
Obiekt klasy \code{flextable} zgodny z APA.
}
\description{
Funkcja przeksztalca wyniki analizy MCDA (TOPSIS, VIKOR, WASPAS, Meta-Ranking)
w sformatowana tabele zgodna ze standardem APA (gotowa do prezentacji i publikacji).
}
