% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabela_apa.R
\name{tabela_apa_html}
\alias{tabela_apa_html}
\title{Generowanie Tabeli HTML w Stylu APA dla SMMRankR}
\usage{
tabela_apa_html(x, tytul = NULL)
}
\arguments{
\item{x}{Obiekt wynikowy z funkcji pakietu (np. \code{topsis_smm_wynik}).}

\item{tytul}{Opcjonalny tytul tabeli.}
}
\value{
Obiekt klasy \code{shiny.tag.list} zawierajacy tabele HTML.
}
\description{
Funkcja przeksztalca wyniki analizy MCDA w sformatowana tabele HTML zgodna ze standardem APA.
}
