% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{uruchom_aplikacje}
\alias{uruchom_aplikacje}
\title{Uruchomienie aplikacji Shiny dla SMMRankR}
\usage{
uruchom_aplikacje(...)
}
\arguments{
\item{...}{Dodatkowe parametry przekazywane do funkcji \code{shiny::runApp}.}
}
\description{
Uruchamia interaktywna aplikacje Shiny sluzaca do analizy wielokryterialnej platform spolecznosciowych.
}
