% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotowanie_danych.R
\name{zbuduj_macierz_decyzyjna}
\alias{zbuduj_macierz_decyzyjna}
\title{Przygotowanie macierzy decyzyjnej SMM}
\usage{
zbuduj_macierz_decyzyjna(
  dane,
  skladnia,
  kolumna_platformy = "Platforma",
  agregacja = mean
)
}
\arguments{
\item{dane}{Ramka danych z surowymi zmiennymi.}

\item{skladnia}{Formula grupujaca zmienne surowe w kryteria kompozytowe (skladnia lavaan).}

\item{kolumna_platformy}{Nazwa kolumny zawierajacej identyfikatory platform (alternatyw). Domyslnie "Platforma".}

\item{agregacja}{Funkcja agregujaca wiersze do poziomu platform. Domyslnie `mean`.}
}
\value{
Macierz decyzyjna (platformy x kryteria) z odpowiednimi nazwami wierszy i kolumn.
}
\description{
Funkcja przetwarza surowe dane marketingowe (np. ankiety, projekty) na sformatowana
macierz decyzyjna na podstawie podanej skladni kryteriow. Czyszczone sa kody bledow (999) i imputowane
braki danych (NA) za pomoca sredniej grupowej dla danej platformy.
}
