% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorytm_mcda.R
\name{oblicz_topsis_smm}
\alias{oblicz_topsis_smm}
\title{Ranking platform SMM metodą TOPSIS}
\usage{
oblicz_topsis_smm(tablica_wynikow, wagi_smm)
}
\arguments{
\item{tablica_wynikow}{Macierz danych (np. TikTok, Instagram) wygenerowana funkcją stworz_macierz_smm.}

\item{wagi_smm}{Wektor wag kryteriów obliczony metodą BWM.}
}
\value{
ranking_koncowy Ramka danych z pozycjami w rankingu.
}
\description{
Funkcja wykonuje pełną procedurę rankingu TOPSIS dla wybranych platform społecznościowych.
}
