% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wizualizacja_rankingu.R
\name{rysuj_ranking_smm}
\alias{rysuj_ranking_smm}
\title{Wykres bąbelkowy wyników rankingu SMM}
\usage{
rysuj_ranking_smm(macierz_danych, wyniki_rankingu)
}
\arguments{
\item{macierz_danych}{Macierz z danymi marketingowymi.}

\item{wyniki_rankingu}{Wyniki obliczone funkcją oblicz_topsis_smm.}
}
\description{
Tworzy mapę strategiczną platform społecznościowych na podstawie wyników MCDA.
}
