% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smm_mcdm_rank.R
\name{smm_rank_score}
\alias{smm_rank_score}
\title{Rankowanie platform i typów treści w mediach społecznościowych}
\usage{
smm_rank_score(scores, weights)
}
\arguments{
\item{scores}{data.frame lub macierz z ocenami kryteriów}

\item{weights}{wektor wag kryteriów}
}
\value{
data.frame z wynikami rankingu
}
\description{
Rankowanie platform i typów treści w mediach społecznościowych
}
